3/30 FollowUp
This is mainly a test to see if my GitHub is still pushing correctly
since last time it game me a funky error, but still went through. But
while I was here I decided to procrastinate again and make more
progress.
TLDR: I fixed some of the problems that Will had.